requirements.txt
azure-storage-blob==2.1.0
azure-storage-common==2.1.0